+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # Unreleased
 - Add `puncons` and `punsafeUncons` functions for deconstructing list.
 
+- Add generic deriving for `PEq`
+
+  Started by [#335](https://github.com/Plutonomicon/plutarch/pull/335)
+
 - `TermCont`: Parametrize by result type; add `MonadFail` instance; etc.
 
   Also, export from `Plutarch.TermCont`, and then from `Plutarch.Prelude` (TermCont is no longer exported by `Plutarch.Internal`).